+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [1.18.3] - 2026-07-14
+
+### Bug Fixes
+
+- fix: handle upload destinations with connectors that have a prefix, hide personal bucket from connector S3 bucket list, set my-assets feature flag to false as default
+
+## [1.18.2] - 2026-07-10
+
+### Bug Fixes
+
+- fix: add SSM_PREFIX environment variable to Assets and Search Lambdas
+
+## [1.18.1] - 2026-07-10
+
+### Bug Fixes
+
+- fix: add object-level authorization to collection api endpoints
+- fix: add SSM_PREFIX to collections API Lambda for CloudFront URL resolution
+
 ## [1.18.0] - 2026-07-08
 
 ### Features